svg/painting/parsing/color-interpolation-invalid.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="800" height="800"><title>SVG Painting: parsing color-interpolation with invalid values</title><metadata><h:link href="https://svgwg.org/svg2-draft/painting.html#ColorInterpolationProperty" rel="help"/><h:meta content="color-interpolation supports only the grammar 'auto | sRGB | linearRGB'." name="assert"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script src="/css/support/parsing-testcommon.js"/><script><![CDATA[

test_invalid_value("color-interpolation", "none");
test_invalid_value("color-interpolation", "auto srgb");

  ]]></script></svg>
